+++ bcipy/task/paradigm/rsvp/copy_phrase.py
@@ -87,7 +87,7 @@ class RSVPCopyPhraseTask(Task):
         'preview_inquiry_key_input', 'preview_inquiry_length',
         'preview_inquiry_progress_method', 'session_file_name',
         'show_feedback', 'show_preview_inquiry', 'spelled_letters_count',
-        'static_trigger_offset', 'stim_color', 'stim_font', 'stim_height',
+        'static_trigger_offset', 'stim_color', 'stim_font', 'stim_height', 'stim_jitter',
         'stim_length', 'stim_number', 'stim_order', 'stim_pos_x', 'stim_pos_y',
         'stim_space_char', 'target_color', 'task_buffer_length', 'task_color',
         'task_font', 'task_height', 'task_text', 'info_pos_x', 'info_pos_y',
@@ -174,7 +174,7 @@ class RSVPCopyPhraseTask(Task):
     def validate_parameters(self) -> None:
         """Validate.
 
-        Confirm Task is configurated with correct parameters and within operating limits.
+        Confirm Task is configured with correct parameters and within operating limits.
         """
 
         # ensure all required parameters are provided
@@ -241,17 +241,8 @@ class RSVPCopyPhraseTask(Task):
             filter_order=self.parameters['filter_order'],
             notch_filter_frequency=self.parameters['notch_filter_frequency'],
             stim_length=self.parameters['stim_length'],
+            stim_jitter=self.parameters['stim_jitter'],
             stim_order=StimuliOrder(self.parameters['stim_order']))
-
-    def await_start(self) -> bool:
-        """Wait on the splash screen for the user to either exit or start."""
-        self.logger.debug('Awaiting user start.')
-        should_continue = get_user_input(
-            self.rsvp,
-            self.parameters['wait_screen_message'],
-            self.parameters['wait_screen_message_color'],
-            first_run=self.first_run)
-        return should_continue
 
     def user_wants_to_continue(self) -> bool:
         """Check if user wants to continue or terminate.
@@ -380,12 +371,10 @@ class RSVPCopyPhraseTask(Task):
         data save location (triggers.txt, session.json)
         """
         self.logger.debug('Starting Copy Phrase Task!')
-        run = self.await_start()
+        run = True
+        self.wait()  # buffer for data processing
 
-        self.wait()  # buffer for data
-
-        while run and self.user_wants_to_continue(
-        ) and self.current_inquiry:
+        while run and self.user_wants_to_continue() and self.current_inquiry:
             target_stimuli = self.next_target()
             stim_times, proceed = self.present_inquiry(
                 self.current_inquiry)
@@ -882,7 +871,7 @@ def _init_copy_phrase_wrapper(min_num_inq, max_num_inq, signal_model, fs, k,
                               stimuli_timing, decision_threshold,
                               backspace_prob, backspace_always_shown,
                               filter_high, filter_low, filter_order,
-                              notch_filter_frequency, stim_length, stim_order):
+                              notch_filter_frequency, stim_length, stim_jitter, stim_order):
     return CopyPhraseWrapper(min_num_inq,
                              max_num_inq,
                              signal_model=signal_model,
@@ -904,4 +893,5 @@ def _init_copy_phrase_wrapper(min_num_inq, max_num_inq, signal_model, fs, k,
                              filter_order=filter_order,
                              notch_filter_frequency=notch_filter_frequency,
                              stim_length=stim_length,
+                             stim_jitter=stim_jitter,
                              stim_order=stim_order)